requirements.txt
google-genai==1.66.0
typer[all]==0.12.5
python-dotenv==1.2.2